LPS-163002 Move package to .search
