+++ modules/apps/dynamic-data-lists/dynamic-data-lists-service/src/main/java/com/liferay/dynamic/data/lists/internal/search/DDLRecordBatchReindexer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.lists.internal.search.util;
+package com.liferay.dynamic.data.lists.internal.search;
 
 /**
  * @author Marcela Cunha
